+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "nodemon src/index.js",
+    "start": "node src/index.js",
+    "dev": "nodemon src/index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "serve": "node src/index.js"
   },
@@ -22,6 +23,7 @@
     "@types/morgan": "^1.9.3",
     "@types/node": "^18.0.3",
     "compression": "^1.7.4",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "errorhandler": "^1.5.1",
     "morgan": "^1.10.0",
