+++ src/Controllers/ProductosController.js
@@ -1,12 +1,7 @@
 import express from "express";
-import morgan from "morgan";
-import compression from "compression";
-import errorhandler from "errorhandler";
-import {ObtenerCategorias, ObtenerProductos} from "../Infraestructura/ProductosService.js";
+import {ObtenerCategorias, ObtenerProductos, ObtenerDetalleProducto} from "../Infraestructura/ProductosService.js";
 
 const productosController = express.Router();
-// app.use(morgan("common"));
-// app.use(compression());
 
 const author = {
     name: "Julian",
@@ -17,7 +12,7 @@ productosController.get('/',(req, res) => {
     res.send("Api tienda ML para productos")
 })
 
-productosController.get("/items", catchUnhandledErrors((req, res) => {
+productosController.get("/api/items", catchUnhandledErrors((req, res) => {
     let promise;
     if (req.query.q && req.query.q.trim()) {
         promise = fetch(`${process.env.API_ML_BUSCADOR_PRODUCTOS}?q=${req.query.q}`).then(response => response.json());
@@ -40,7 +35,7 @@ productosController.get("/items", catchUnhandledErrors((req, res) => {
         });
 }));
 
-productosController.get("/items/:id", catchUnhandledErrors((req, res) => {
+productosController.get("/api/items/:id", catchUnhandledErrors((req, res) => {
     return fetch(`${process.env.API_ML_PRODUCTOS_INFO}/items/${req.params.id}`)
         .then(async response => {
             if (response.status !== 200) {
@@ -58,22 +53,9 @@ productosController.get("/items/:id", catchUnhandledErrors((req, res) => {
                         .then(([ category, description ]) => ([ article, category, description ]));
                     })
                     .then(([ article, category, description ]) => {
-                        let categories = (category.path_from_root || [])
-                            .map(path => path.name);
-                        let item = {
-                            id: article.id,
-                            title: article.title,
-                            price: {
-                                currency: article.currency_id,
-                                amount: article.available_quantity,
-                                decimals: article.price
-                            },
-                            picture: article.pictures && article.pictures.length > 0 ? article.pictures[0].url : article.thumbnail,
-                            condition: article.condition,
-                            free_shipping: (article.shipping && article.shipping.free_shipping === true),
-                            sold_quantity: article.sold_quantity,
-                            description: description.plain_text
-                        };
+                        let detalleProducto = ObtenerDetalleProducto(category, article, description);   
+                        let categories = detalleProducto.categories; 
+                        let item = detalleProducto.item                 
                         res.send({
                             author,
                             categories,
